update to be compatible with bitcoinj 0.16.1
genesis block initialization was moved to lazy getter in bitcoinj, it could be done here too, but keeping the change small by just adding the field like it was before

alertSigningKey was removed in bitcoinj, removing here too, even though some forks might use it

+++ core/src/main/java/org/bitcoinj/core/MerkleBranch.java
@@ -96,7 +96,7 @@ public class MerkleBranch extends ChildMessage {
     protected void parse() throws ProtocolException {
         cursor = offset;
 
-        final int hashCount = (int) readVarInt();
+        final int hashCount = (int) readVarInt().longValue();
         optimalEncodingMessageSize += VarInt.sizeOf(hashCount);
         hashes = new ArrayList<Sha256Hash>(hashCount);
         for (int hashIdx = 0; hashIdx < hashCount; hashIdx++) {
